Converted infinite rays into a vision cone,
I saw something similar and i thought that having a limited range will be more realistic for the resulting NN, It also provides a cool visual that was some maths to figure out. 
Also redid the controls to more correctly move around the screen. However this needs some work for the NN to take control.

And also added in collisions so the particle can nolonger go through walls. 

This commit really should have been 4. 

1 collisions
2 controls 
3 visual cone
4 highlighting seen walls.